add isometric following camera and isometric player movement, looking at cursor
Camera now is not child of player. Created a script for the camera to follow the player.
The ability of the object to look at the cursor implement by calculating direction between game object position on Screen and mouse position on Screen and convert it to world position from camera local space